pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "openstreetmap-ng"
packages = [
  {include = "src"},
]
version = "0.0.0"

[tool.poetry.dependencies]
aioboto3 = "^12.1.0"
aiosmtplib = "^3.0.1"
alembic = "^1.13.1"
anyio = "^4.2.0"
argon2-cffi = "^23.1.0"
arrow = "^1.3.0"
asyncache = "^0.3.1"
asyncpg = "^0.29.0"
authlib = "^1.3.0"
babel = "^2.14.0"
beautifulsoup4 = "^4.12.2"
bleach = "^6.1.0"
cachetools = "^5.3.2"
cryptography = "^41.0.7"
cython = "^3.0.7"
email-validator = "^2.1.0.post1"
fastapi = "^0.108.0"
feedgen = "^1.0.0"
geoalchemy2 = "^0.14.3"
httpx = {extras = ["brotli", "http2"], version = "^0.26.0"}
humanize = "^4.9.0"
jinja2 = "^3.1.2"
lxml = "^5.0.1"
markdown-it-py = {extras = ["plugins"], version = "^3.0.0"}
msgspec = "^0.18.5"
networkx = "^3.2.1"
numpy = "^1.26.3"
orjson = "^3.9.10"
phonenumbers = "^8.13.27"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
pycryptodome = "^3.19.1"
pydantic = "^2.5.3"
pyinstrument = "^4.6.1"
pyproj = "^3.6.1"
python = "^3.12"
python-dateutil = "^2.8.2"
python-magic = "^0.4.27"
python-multipart = "^0.0.6"
pytz = "^2023.3.post1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
rfc3986 = "^2.0.0"
scikit-learn = "^1.3.2"
setuptools = "^69.0.3"
shapely = "^2.0.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
tqdm = "^4.66.1"
uvicorn = {extras = ["standard"], version = "^0.25.0"}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
xmltodict = "^0.13.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core", "setuptools", "wheel", "Cython"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "src/alembic/versions",
]

indent-width = 4
line-length = 120
target-version = "py312"

[tool.ruff.lint]
ignore = [
  "S101", # assert
  "PT011", # pytest-raises-too-broad
]
select = [
  # see https://docs.astral.sh/ruff/rules/ for rules documentation
  "A", # flake8-builtins
  "ARG", # flake8-unused-arguments
  "ASYNC", # flake8-async
  "B", # flake8-bugbear
  # "COM", # flake8-commas
  "C4", # flake8-comprehensions
  "E4", # pycodestyle
  "E7",
  "E9",
  "F", # pyflakes
  # "FBT", # flake8-boolean-trap
  "FLY", # flynt
  # "FURB", # refurb (preview)
  "G", # flake8-logging-format
  "I", # isort
  "INT", # flake8-gettext
  # "LOG", # flake8-logging (preview)
  "N", # pep8-naming
  "NPY", # numpy
  "UP", # pyupgrade
  "Q", # flake8-quotes
  "PERF", # perflint
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  # "PL", # pylint
  "PT", # flake8-pytest-style
  "PTH", # flake8-use-pathlib
  "PYI", # flake8-pyi
  "RSE", # flake8-raise
  "RUF", # ruff
  "S", # flake8-bandit
  "SIM", # flake8-simplify
  "SLF", # flake8-self
  "SLOT", # flake8-slots
  "T10", # flake8-debugger
  # "T20", # flake8-print
  # "TRY", # tryceratops
  "YTT", # flake8-2020
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "lf"

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id", "input", "open", "type"]

[tool.ruff.lint.flake8-gettext]
extend-function-names = ["pgettext", "npgettext"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["ET"]

[tool.ruff.lint.pylint]
max-args = 10
